making the prefix fully configurable and moving some functions from index.js to profile.js in order to get a convenient and simple code at first glance
