+++ services/config.js
@@ -5,7 +5,7 @@ export default
 {
 
     "prefix": "!",
-    "api_key": "RGAPI-5a2e3fa9-62d4-459c-b856-d9263b612bc6",
+    "api_key": "RGAPI-5e43791d-4906-4ea3-b970-2cbf67639b85",
     "bot_key": "NzIwNzYwNDczNjAzMDE0Nzk5.XuKq1A.jJRAUKl35lMUu24DTC-aFdGKnkA"
 
 }
